api/requirements.txt
asgiref==3.8.1
astroid==3.3.9
click==8.1.8
dill==0.4.0
Django==5.2
djangorestframework==3.16.0
isort==6.0.1
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.7
python-dotenv==1.0.1
sqlparse==0.5.3
tomlkit==0.13.2
